% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comb_LAD.R
\name{comb_LAD}
\alias{comb_LAD}
\title{Least Absolute Deviation Forecast Combination}
\usage{
comb_LAD(x)
}
\arguments{
\item{x}{An object of class 'foreccomb'. Contains training set (actual values + matrix of model forecasts) and optionally a test set.}
}
\value{
Returns an object of class 'foreccomb_res'
\itemize{
\item Method Returns the used forecast combination method.
\item Models Returns the individual input models that were used for the forecast combinations.
\item Weights Returns the combination weights obtained by applying the combination method to the training set.
\item Intercept Returns the intercept of the model.
\item Fitted Returns the fitted values of the combination method for the training set.
\item Accuracy_Train Returns range of summary measures of the forecast accuracy for the training set.
\item Forecasts_Test Returns forecasts produced by the combination method for the test set. Only returned if input included a forecast matrix for the test set.
\item Accuracy_Test Returns range of summary measures of the forecast accuracy for the test set. Only returned if input included a forecast matrix and a vector of actual values for the test set.
\item Input_Data Returns the data forwarded to the method.
}
}
\description{
Computes least absolute deviation forecast combination weights via the \emph{ForecastCombinations} package.
}
\details{
The function \code{comb_LAD} is a wrapper around the least absolute deviation (LAD) forecast combination implementation of the
\emph{ForecastCombinations} package.

The defining property of \code{comb_LAD} is that it does not minimize the squared error loss like \code{\link{comb_OLS}} and
\code{\link{comb_CLS}}, but the absolute values of the errors. This has the advantage that it is more robust against outliers,
that is, in the case of forecast combination, \code{comb_LAD} tends to penalize models, which have for some observations high errors,
less harshly than the least squares methods would.

The results are stored in an object of class 'foreccomb_res', for which separate plot and summary functions are provided.
}
\examples{
obs <- rnorm(100)
preds <- matrix(rnorm(1000, 1), 100, 10)
train_o<-obs[1:80]
train_p<-preds[1:80,]
test_o<-obs[81:100]
test_p<-preds[81:100,]

data<-foreccomb(train_o, train_p, test_o, test_p)
comb_LAD(data)

}
\seealso{
\code{\link[GeomComb]{foreccomb}},
\code{\link[GeomComb]{plot.foreccomb_res}},
\code{\link[GeomComb]{summary.foreccomb_res}},
\code{\link[forecast]{accuracy}}
}
\keyword{ts}

