% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comb_OLS.R
\name{comb_OLS}
\alias{comb_OLS}
\title{Ordinary Least Squares Forecast Combination}
\usage{
comb_OLS(x)
}
\arguments{
\item{x}{An object of class 'foreccomb'. Contains training set (actual values + matrix of model forecasts) and optionally a test set.}
}
\value{
Returns an object of class 'foreccomb_res'
\itemize{
\item Method Returns the used forecast combination method.
\item Models Returns the individual input models that were used for the forecast combinations.
\item Weights Returns the combination weights obtained by applying the combination method to the training set.
\item Intercept Returns the intercept of the linear regression.
\item Fitted Returns the fitted values of the combination method for the training set.
\item Accuracy_Train Returns range of summary measures of the forecast accuracy for the training set.
\item Forecasts_Test Returns forecasts produced by the combination method for the test set. Only returned if input included a forecast matrix for the test set.
\item Accuracy_Test Returns range of summary measures of the forecast accuracy for the test set. Only returned if input included a forecast matrix and a vector of actual values for the test set.
}
}
\description{
Computes forecast combination weights via the \emph{ForecastCombinations} package.
}
\details{
The function \code{comb_ols} is a wrapper around the ordinary least squares (OLS) forecast combination implementation of the
\emph{ForecastCombinations} package.
 
The OLS combination method (Granger and Ramanathan (1984)) uses ordinary least squares to
estimate the weights, \eqn{\mathbf{w}^{OLS} = (w_1, \ldots, w_N)'}, as well as an intercept, \eqn{b}, for the combination of
the forecasts.

Suppose that there are N different predictors  \eqn{\mathbf{f}_t = (f_{1t}, \ldots, f_{Nt})'},
then the forecast combination for one data point can be represented as:
\deqn{y_t = b + \sum_{i=1}^{N} w_i f_i}  
The results are stored in an object of class 'foreccomb_res', for which separate plot and summary functions are provided.
}
\examples{
obs <- rnorm(100)
preds <- matrix(rnorm(1000, 1), 100, 10)
train_o<-obs[1:80]
train_p<-preds[1:80,]
test_o<-obs[81:100]
test_p<-preds[81:100,]

data<-foreccomb(train_o, train_p, test_o, test_p)
comb_OLS(data)

}
\references{
Granger, C., and Ramanathan, R. (1984). Improved Methods Of Combining Forecasts. \emph{Journal of Forecasting}, \bold{3(2)}, 97--204.
}
\seealso{
\code{\link[GeomComb]{foreccomb}},
\code{\link[GeomComb]{plot.foreccomb_res}},
\code{\link[GeomComb]{summary.foreccomb_res}},
\code{\link[forecast]{accuracy}}
}
\keyword{ts}

