% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_dispersion.R
\name{cs_dispersion}
\alias{cs_dispersion}
\title{Compute Cross-Sectional Dispersion}
\usage{
cs_dispersion(x, measure = NULL, plot = FALSE)
}
\arguments{
\item{x}{An object of class 'foreccomb'. Contains training set (actual values + matrix of model forecasts) and optionally a test set.}

\item{measure}{Cross-sectional dispersion measure, one of: "SD" = standard deviation; "IQR" = interquartile range; or "Range" = range.}

\item{plot}{logical. If TRUE, evolution of cross-sectional forecast dispersion in plotted as ggplot.}
}
\value{
Returns a vector of the evolution of cross-sectional dispersion over the sample period (using the selected dispersion measure)
}
\description{
Computes (time-varying) dispersion measures for the cross section of individual model forecasts that are the input of forecast combination.
}
\details{
Previous research in the forecast combination literature has documented that regression-based combination methods tend to have relative advantage when one or more individual model forecasts are better than the rest, while eigenvector-based methods tend to have relative advantage when individual model forecasts are in the same ball park.
}
\examples{
obs <- rnorm(100)
preds <- matrix(rnorm(1000, 1), 100, 10)
train_o<-obs[1:80]
train_p<-preds[1:80,]
test_o<-obs[81:100]
test_p<-preds[81:100,]

data<-foreccomb(train_o, train_p, test_o, test_p)
cs_dispersion(data, measure = "IQR")

}
\references{
Adamson, K. A., and Prion S. (2013). Making Sense of Methods and Measurement: Measures of Variability. \emph{Clinical Simulation in Nusing}, \bold{9}, e559--e560.

Hsiao, C., and Wan, S. K. (2014). Is There An Optimal Forecast Combination? \emph{Journal of Econometrics}, \bold{178(2)}, 294--309.
}
\seealso{
\code{\link[GeomComb]{foreccomb}}
}
\keyword{ts}

