% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foreccomb.R
\name{foreccomb}
\alias{foreccomb}
\title{Create Forecast Combination Input Data}
\usage{
foreccomb(observed_vector, prediction_matrix, newobs = NULL,
  newpreds = NULL, byrow = FALSE, na.impute = TRUE, criterion = "RMSE")
}
\arguments{
\item{observed_vector}{A vector or univariate time series; contains 'actual values' for training set.}

\item{prediction_matrix}{A matrix or multivariate time series; contains individual model forecasts for training set.}

\item{newobs}{A vector or univariate time series; contains 'actual values' if a test set is used.}

\item{newpreds}{A matrix or multivariate time series; contains individual model forecasts if a test set is used. Does not
require specification of 'newobs' -- in the case in which a forecaster only wants to train the forecast combination method
with a training set and apply it to future individual model forecasts, only 'newpreds' is required, not 'newobs'.}

\item{byrow}{logical. The default (FALSE) assumes that each column of the forecast matrices ('prediction_matrix' and -- if
specified -- 'newpreds') contains forecasts from one forecast model; if each row of the matrices contains forecasts from
one forecast model, set to TRUE.}

\item{na.impute}{logical. The default (TRUE) behavior is to impute missing values via the cross-validated spline approach of
the mtsdi package. If set to FALSE, forecasts with missing values will be removed. Missing values in the observed data are never
imputed}

\item{criterion}{One of "RMSE", "MAE", or "MAPE".}
}
\value{
Returns an object of class 'foreccomb'
\itemize{
 \item Actual_Train Actual Values (Training Set).
 \item Forecasts_Train Retained (non-missing) cross-section of individual model forecasts (Training Set).
 \item Actual_Test Actual Values (Test Set). If 'newobs' was provided.
 \item Forecasts_Test Retained (non-missing) cross-section of individual model forecasts (Test Set). If 'newpreds' was provided.
 \item nmodels Number of retained (non-missing) individual model forecasts.
 \item modelnames Model Names of the retained individual forecast models -- either provided or generic names are created (see above).
}
}
\description{
Structures cross-sectional input data (individual model forecasts) for forecast combination. Stores data as S3 class 'foreccomb' that is passed on to the geometric forecast combination techniques.
}
\details{
The function imports the column names of the prediction matrix (if byrow = FALSE, otherwise the row names) as model names; #
if no column names are specified, generic model names are created.
}
\examples{
obs <- rnorm(100)
preds <- matrix(rnorm(1000, 1), 100, 10)
train_o<-obs[1:80]
train_p<-preds[1:80,]
test_o<-obs[81:100]
test_p<-preds[81:100,]

## Example with a training set only:
foreccomb(train_o, train_p)

## Example with a training set and future individual forecasts:
foreccomb(train_o, train_p, newpreds=test_p)

## Example with a training set and a full test set:
foreccomb(train_o, train_p, test_o, test_p)

## Example with individual forecast models being stored in rows:
preds <- matrix(rnorm(1000, 1), 10, 100)
train_p <- preds[,1:80]
foreccomb(train_o, train_p, byrow = TRUE)

}
\references{
~put references to the literature/web site here ~
}
\seealso{
~~objects to See Also as \code{\link{help}}, ~~~
}
\keyword{ts}

