% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foreccomb_res.R
\name{foreccomb_res}
\alias{foreccomb_res}
\title{Result Rbject for the Forecast Combination Methods}
\usage{
foreccomb_res(method, modelnames, fitted, accuracy_insample, input_data,
  predict = NULL, intercept = NULL, weights = NULL, pred = NULL,
  accuracy_outsample = NULL, trim_factor = NULL, top_predictors = NULL,
  ranking = NULL)
}
\arguments{
\item{method}{Name of the method.}

\item{modelnames}{Names of the models provided by the user.}

\item{fitted}{The fitted values.}

\item{accuracy_insample}{Insample accuracy obtained by the method.}

\item{input_data}{Contains the input data provided by the user.}

\item{predict}{(optional) Function used to conduct predictions for new forecasts.}

\item{intercept}{(optional) Intercept of the method, if it requires one.}

\item{weights}{(optional) Weights of the method, if it requires one.}

\item{pred}{= (optional) Predictions on a test set.}

\item{accuracy_outsample}{= (optional) Accuracy on the test set.}

\item{trim_factor}{= (optional) Trim factor used in some of the methods.}

\item{top_predictors}{(optional) Number of retained predictors.}

\item{ranking}{(optional) Ranking of the predictors that determines which models are removed in the trimming step.}
}
\value{
Returns an object of class \code{foreccomb_res}.
}
\description{
Stores the results and inputs for some combined forecasts.
}
\author{
Christoph E. Weiss, Gernot R. Roetzer
}

