% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.foreccomb_res.R
\name{print.foreccomb_res_summary}
\alias{print.foreccomb_res_summary}
\title{PLACEHOLDER for print.foreccomb_res_summary}
\usage{
\method{print}{foreccomb_res_summary}(object, ...)
}
\arguments{
\item{object}{An object of class 'foreccomb'. Contains training set (actual values + matrix of model forecasts) and optionally a test set.}

\item{...}{Additional parameters}
}
\value{
Returns formatted output of the foreccomb_res_summary object
}
\description{
Computes forecast combination weights according to the standard eigenvector approach by Hsiao and Wan (2014) and produces forecasts for the test set, if provided.
}
\details{
The standard eigenvector approach retrieves combination weights from the sample estimated mean squared prediction error matrix 
as follows: Suppose \eqn{y_t} is a variable of interest, there are N not perfectly collinear predictors, 
\eqn{\bold{f}_t = (f_{1t}, \ldots, f_{Nt})'}, \eqn{\Sigma} is the (positive definite) 
mean squared prediction error matrix of \eqn{\bold{f}_t} and \eqn{\bold{e}} is an \eqn{N \times 1}{N * 1} vector of \eqn{(1,\ldots,1)'}. 
The N positive eigenvalues are then arranged in increasing order \eqn{(\Phi_1 = \Phi_{min}, \Phi_2, \ldots, \Phi_N)}, and \eqn{\bold{w^j}} 
is defined as the eigenvector corresponding to \eqn{\Phi_j}. The combination weights \eqn{\bold{w} = (w_1,\ldots,w_N)'} are then 
chosen corresponding to the minimum of \eqn{\left{ \frac{\Phi_1}{d_1^2}, \frac{\Phi_2}{d_2^2},\ldots,\frac{\Phi_N}{d_N^2}\right}}, 
denoted as \eqn{\bold{w}^l}, where \eqn{d_j = \bold{e}'\bold{w}^j}, as:
\deqn{\bold{w}^{EIG1} = \frac{1}{d_l} \bold{w}^l}
The results are stored in an object of class 'foreccomb_res', for which separate plot and summary functions are provided.
}

