% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge.R
\name{ridge}
\alias{ridge}
\title{Ridge regression}
\usage{
ridge(x, y, lambda = 10^seq(-10, 10, length.out = 100))
}
\arguments{
\item{x}{matrix of predictors}

\item{y}{response variable}

\item{lambda}{vector of ridge parameters}
}
\value{
An object of class \code{ridge} with the following components:
\item{coef}{matrix of coefficients}
\item{ym}{mean of the response variable}
\item{xm}{mean of the predictors}    
\item{xsd}{standard deviation of the predictors}
\item{lambda}{vector of ridge parameters}
\item{best_lam}{best ridge parameter according to GCV}
\item{fitted_values}{fitted values}
\item{residuals}{residuals}
\item{GCV}{Generalized Cross Validation}
\item{BIC}{Bayesian Information Criterion}
\item{x}{predictors}
\item{y}{response variable}
}
\description{
Fit a ridge regression model.
}
\examples{

x <- matrix(rnorm(100*10), 100, 10)
y <- rnorm(100)
ridge_model <- ridge(x, y)
predict(ridge_model, x)

}
