% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.foreccomb_res.R
\name{summary.foreccomb_res}
\alias{print.foreccomb_res_summary}
\alias{summary.foreccomb_res}
\title{Summary of Forecast Combination}
\usage{
\method{summary}{foreccomb_res}(object, ...)

\method{print}{foreccomb_res_summary}(x, ...)
}
\arguments{
\item{object}{An object of class 'foreccomb'. Contains training set (actual values + matrix of model forecasts) and optionally a test set.}

\item{...}{Additional parameters}

\item{x}{An object of class 'foreccomb'. Contains training set (actual values + matrix of model forecasts) and optionally a test set.}
}
\value{
PLACEHOLDER Returns an object of class 'summary.foreccomb_res'
\itemize{
 \item Actual_Train Actual Values (Training Set).
 \item Forecasts_Train Retained (non-missing) cross-section of individual model forecasts (Training Set).
 \item Actual_Test Actual Values (Test Set). If 'newobs' was provided.
 \item Forecasts_Test Retained (non-missing) cross-section of individual model forecasts (Test Set). If 'newpreds' was provided.
 \item nmodels Number of retained (non-missing) individual model forecasts.
 \item modelnames Model Names of the retained individual forecast models -- either provided or generic names are created (see above).
}
}
\description{
Computes forecast combination weights according to the standard eigenvector approach by Hsiao and Wan (2014) and produces forecasts for the test set, if provided.
}
\details{
The standard eigenvector approach retrieves combination weights from the sample estimated mean squared prediction error matrix
as follows:
The results are stored in an object of class 'foreccomb_res', for which separate plot and summary functions are provided.
}
\examples{
obs <- rnorm(100)
preds <- matrix(rnorm(1000, 1), 100, 10)
train_o<-obs[1:80]
train_p<-preds[1:80,]
test_o<-obs[81:100]
test_p<-preds[81:100,]

data<-foreccomb(train_o, train_p, test_o, test_p)
ev_comb_EIG1(data)

}
\references{
Hsiao, C., and Wan, S. K. (2014). Is There An Optimal Forecast Combination? \emph{Journal of Econometrics}, \bold{178(2)}, 294--309.
}
\seealso{
\code{\link[GeomComb]{foreccomb}},
\code{\link[GeomComb]{plot.foreccomb_res}},
\code{\link[GeomComb]{summary.foreccomb_res}},
\code{\link[forecast]{accuracy}}
}
\keyword{ts}

